LPS-89499 portal-search-elasticsearch6-impl: move legacy query translators
https://issues.liferay.com/browse/LPS-89499
