+++ modules/apps/portal-search-elasticsearch6/portal-search-elasticsearch6-impl/src/main/java/com/liferay/portal/search/elasticsearch6/internal/legacy/query/MatchAllQueryTranslatorImpl.java
@@ -12,32 +12,32 @@
  * details.
  */
 
-package com.liferay.portal.search.elasticsearch6.internal.query;
+package com.liferay.portal.search.elasticsearch6.internal.legacy.query;
 
-import com.liferay.portal.kernel.search.generic.StringQuery;
+import com.liferay.portal.kernel.search.generic.MatchAllQuery;
 
+import org.elasticsearch.index.query.MatchAllQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.index.query.QueryStringQueryBuilder;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Michael C. Han
  */
-@Component(immediate = true, service = StringQueryTranslator.class)
-public class StringQueryTranslatorImpl implements StringQueryTranslator {
+@Component(immediate = true, service = MatchAllQueryTranslator.class)
+public class MatchAllQueryTranslatorImpl implements MatchAllQueryTranslator {
 
 	@Override
-	public QueryBuilder translate(StringQuery stringQuery) {
-		QueryStringQueryBuilder queryStringQueryBuilder =
-			QueryBuilders.queryStringQuery(stringQuery.getQuery());
+	public QueryBuilder translate(MatchAllQuery matchAllQuery) {
+		MatchAllQueryBuilder matchAllQueryBuilder =
+			QueryBuilders.matchAllQuery();
 
-		if (!stringQuery.isDefaultBoost()) {
-			queryStringQueryBuilder.boost(stringQuery.getBoost());
+		if (!matchAllQuery.isDefaultBoost()) {
+			matchAllQueryBuilder.boost(matchAllQuery.getBoost());
 		}
 
-		return queryStringQueryBuilder;
+		return matchAllQueryBuilder;
 	}
 
 }
