+++ modules/apps/portal-search-elasticsearch6/portal-search-elasticsearch6-impl/src/main/java/com/liferay/portal/search/elasticsearch6/internal/legacy/query/NestedQueryTranslatorImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.search.elasticsearch6.internal.query;
+package com.liferay.portal.search.elasticsearch6.internal.legacy.query;
 
 import com.liferay.portal.kernel.search.Query;
 import com.liferay.portal.kernel.search.generic.NestedQuery;
