+++ modules/apps/portal-search-elasticsearch6/portal-search-elasticsearch6-impl/src/main/java/com/liferay/portal/search/elasticsearch6/internal/legacy/query/TermQueryTranslatorImpl.java
@@ -12,35 +12,39 @@
  * details.
  */
 
-package com.liferay.portal.search.elasticsearch6.internal.query;
+package com.liferay.portal.search.elasticsearch6.internal.legacy.query;
 
 import com.liferay.portal.kernel.search.QueryTerm;
-import com.liferay.portal.kernel.search.WildcardQuery;
+import com.liferay.portal.kernel.search.TermQuery;
 
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.index.query.WildcardQueryBuilder;
+import org.elasticsearch.index.query.TermQueryBuilder;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
+ * @author André de Oliveira
  * @author Miguel Angelo Caldas Gallindo
  */
-@Component(immediate = true, service = WildcardQueryTranslator.class)
-public class WildcardQueryTranslatorImpl implements WildcardQueryTranslator {
+@Component(immediate = true, service = TermQueryTranslator.class)
+public class TermQueryTranslatorImpl implements TermQueryTranslator {
 
 	@Override
-	public QueryBuilder translate(WildcardQuery wildcardQuery) {
-		QueryTerm queryTerm = wildcardQuery.getQueryTerm();
+	public QueryBuilder translate(TermQuery termQuery) {
+		QueryTerm queryTerm = termQuery.getQueryTerm();
 
-		WildcardQueryBuilder wildcardQueryBuilder = QueryBuilders.wildcardQuery(
-			queryTerm.getField(), queryTerm.getValue());
+		String field = queryTerm.getField();
+		String value = queryTerm.getValue();
 
-		if (!wildcardQuery.isDefaultBoost()) {
-			wildcardQueryBuilder.boost(wildcardQuery.getBoost());
+		TermQueryBuilder termQueryBuilder = QueryBuilders.termQuery(
+			field, value);
+
+		if (!termQuery.isDefaultBoost()) {
+			termQueryBuilder.boost(termQuery.getBoost());
 		}
 
-		return wildcardQueryBuilder;
+		return termQueryBuilder;
 	}
 
 }
